moved from no.uio.ifi to no.elixir.fega and updated version to 2.0.0
