+++ pom.xml
@@ -8,9 +8,9 @@
         <version>3.1.3</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
-    <groupId>no.uio.ifi</groupId>
+    <groupId>no.elixir.fega</groupId>
     <artifactId>localega-tsd-proxy</artifactId>
-    <version>1.1.7</version>
+    <version>2.0.0</version>
     <name>localega-tsd-proxy</name>
     <description>LocalEGA TSD Proxy</description>
