+++ src/main/java/no/elixir/fega/ltp/aspects/ProcessArgumentsAspect.java
@@ -1,4 +1,4 @@
-package no.uio.ifi.ltp.aspects;
+package no.elixir.fega.ltp.aspects;
 
 import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.ProceedingJoinPoint;
@@ -42,7 +42,7 @@ public class ProcessArgumentsAspect {
      * @throws Throwable In case of error.
      */
     @SuppressWarnings("rawtypes")
-    @Around("execution(public * no.uio.ifi.ltp.controllers.rest.ProxyController.stream(..)) || execution(public * no.uio.ifi.ltp.controllers.rest.ProxyController.deleteFile(..))")
+    @Around("execution(public * no.elixir.fega.ltp.controllers.rest.ProxyController.stream(..)) || execution(public * no.elixir.fega.ltp.controllers.rest.ProxyController.deleteFile(..))")
     public Object processArguments(ProceedingJoinPoint joinPoint) throws Throwable {
         try {
             Object[] arguments = joinPoint.getArgs();
