+++ src/main/java/no/elixir/fega/ltp/dto/Credentials.java
@@ -1,4 +1,4 @@
-package no.uio.ifi.ltp.dto;
+package no.elixir.fega.ltp.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
