+++ src/main/java/no/elixir/fega/ltp/dto/EncryptedIntegrity.java
@@ -1,4 +1,4 @@
-package no.uio.ifi.ltp.dto;
+package no.elixir.fega.ltp.dto;
 
 import com.google.gson.annotations.SerializedName;
 import lombok.Data;
