+++ src/test/java/no/elixir/fega/ltp/LocalEGATSDProxyApplicationTests.java
@@ -1,4 +1,4 @@
-package no.uio.ifi.ltp;
+package no.elixir.fega.ltp;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
